fix: resolve CI failures for dead code and binary discovery
- Remove unreachable maybe_strip_colors #[cfg(not(feature = "display"))]
  variant that caused dead_code error on wasm32 builds
- Add shared test_utils module with prqlc_command() that auto-builds
  the prqlc binary on demand using std::sync::Once, fixing test failures
  when cargo test doesn't build the binary (e.g. minimal-versions check)
