+++ prqlc/prqlc-cli/src/cli/docs_generator.rs
@@ -310,10 +310,9 @@ Generated with [prqlc](https://prql-lang.org/) {}.
 
 #[cfg(test)]
 mod tests {
-    use std::process::Command;
-
     use insta_cmd::assert_cmd_snapshot;
-    use insta_cmd::get_cargo_bin;
+
+    use super::super::test_utils::prqlc_command;
 
     #[test]
     fn generate_html_docs() {
@@ -507,24 +506,5 @@ mod tests {
 
         ----- stderr -----
         ");
-    }
-
-    fn prqlc_command() -> Command {
-        let mut cmd = Command::new(get_cargo_bin("prqlc"));
-        normalize_prqlc(&mut cmd);
-        cmd
-    }
-
-    fn normalize_prqlc(cmd: &mut Command) -> &mut Command {
-        cmd
-            // We set `CLICOLOR_FORCE` in CI to force color output, but we don't want `prqlc` to
-            // output color for our snapshot tests. And it seems to override the
-            // `--color=never` flag.
-            .env_remove("CLICOLOR_FORCE")
-            .env("NO_COLOR", "1")
-            .args(["--color=never"])
-            // We don't want the tests to be affected by the user's `RUST_BACKTRACE` setting.
-            .env_remove("RUST_BACKTRACE")
-            .env_remove("RUST_LOG")
     }
 }
